Internal function to convert sprawlext to sfc_POLYGON
This was necessary since the method to convert sprawlext to sfc_POLYGON and sfc_POINT was not recognised by running the package. Now the method calls an internal function(`.sprawlext_to_poly()`), which is used within other functions instead of `as(sprawlext_object, "sfc_POLYGON")`. The same for `.sprawlext_to_point()` instead of `as(sprawlext_object, "sfc_POINT")`.